+++ pyccl/baryons/schneider15.py
@@ -7,26 +7,21 @@ from . import Baryons
 
 
 class BaryonsSchneider15(Baryons):
-    """The "BCM" model boost factor for baryons.
+    """The "BCM" model boost factor for baryons. BCM stands for the
+    "baryonic correction model" of `Schneider & Teyssier 2015
+    <https://arxiv.org/abs/1510.06034>`_. See the
 
-    .. note:: BCM stands for the "baryonic correction model" of `Schneider &
-              Teyssier 2015 <https://arxiv.org/abs/1510.06034>`_. See the
-              `DESC Note <https://github.com/LSSTDESC/CCL/blob/master/doc\
-/0000-ccl_note/main.pdf>`_
-              for details.
-
-              The boost factor is applied multiplicatively so that
-              :math:`P_{\\rm bar.}(k, a) = P_{\\rm DMO}(k, a)\\,
-              f_{\\rm BCM}(k, a)`.
+    The boost factor is applied multiplicatively so that
+    :math:`P_{\\rm bar.}(k, a) = P_{\\rm DMO}(k, a)\\, f_{\\rm BCM}(k, a)`.
 
     Args:
         log10Mc (:obj:`float`): logarithmic mass scale of hot
             gas suppression. Defaults to
             :math:`\\log_{10}(1.2\\,10^{14}\\,M_\\odot)`.
-        eta_b (:obj:`float`): ratio of escape to ejection radii (see
-            Teyssier et al. 2015). Defaults to 0.5.
+        eta_b (:obj:`float`): ratio of escape to ejection radii.
+            Defaults to 0.5.
         k_s (:obj:`float`): Characteristic scale (wavenumber) of
-            the stellar component in units of :math:`{\\rm Mpc}\\,h^{-1}`.
+            the stellar component in units of :math:`h\\,{\\rm Mpc}^{-1}`.
             Defaults to 55.
     """
     name = 'Schneider15'
@@ -46,8 +41,8 @@ class BaryonsSchneider15(Baryons):
             a (:obj:`float` or `array`): Scale factor.
 
         Returns:
-            :obj:`float` or `array`: Correction factor to apply to
-                the power spectrum.
+            :obj:`float` or `array`: Correction factor to apply to \
+            the power spectrum.
         """ # noqa
         a_use, k_use = map(np.atleast_1d, [a, k])
         a_use, k_use = a_use[:, None], k_use[None, :]