+++ pyccl/modified_gravity/modified_gravity_base.py
@@ -1,0 +1,8 @@
+__all__ = ("ModifiedGravity",)
+
+
+class ModifiedGravity:
+    """Base container class for modified gravity
+    parametrisations.
+    """
+    parametrization: str = None